+++ fabric/src/main/java/io/github/sakurawald/module/initializer/head/config/model/HeadConfigModel.java
@@ -1,10 +1,16 @@
 package io.github.sakurawald.module.initializer.head.config.model;
 
+import com.google.gson.annotations.SerializedName;
 import io.github.sakurawald.module.initializer.head.structure.EconomyType;
 import org.jetbrains.annotations.NotNull;
 
 public class HeadConfigModel {
+
     public EconomyType economy_type = EconomyType.ITEM;
-    public @NotNull String cost_type = "minecraft:emerald_block";
-    public int cost_amount = 1;
+
+    @SerializedName(value = "cost_item_type", alternate = "cost_type")
+    public @NotNull String cost_item_type = "minecraft:emerald_block";
+
+    @SerializedName(value = "cost_item_amount", alternate = "cost_amount")
+    public int cost_item_amount = 1;
 }