+++ fabric/src/main/java/io/github/sakurawald/module/initializer/head/gui/SearchHeadsInputGui.java
@@ -30,11 +30,14 @@ public class SearchHeadsInputGui extends InputSignGui {
         }
 
         /* Filter the entities by keywords. */
-        List<Head> entities = HeadProvider.getLoadedHeads().values().stream()
+        List<Head> entities = HeadProvider.getLoadedHeads().values()
+            .stream()
             .filter(head -> head.name.toLowerCase().contains(keywords.toLowerCase())
                 || head.getTagsOrEmpty().toLowerCase().contains(keywords.toLowerCase()))
             .collect(Collectors.toList());
+
         Text title = TextHelper.getTextByKey(player, "gui.search.title", keywords);
-        new CategoryHeadGui(this.parentGui, player, title, entities, 0).open();
+        new CategoryHeadsGui(this.parentGui, player, title, entities, 0)
+            .open();
     }
 }